refactor: 스타일드 컴포넌트 파일 분리
