+++ src/components/board/BulletinCard/BulletinCard.tsx
@@ -1,63 +1,17 @@
-import styled from "styled-components";
-import tw from "tailwind-styled-components";
+import HeartBtn from "@/components/board/Button/HeartBtn/HeartBtn";
 
-import HeartBtn from "@/components/board/HeartBtn";
-
-const Card = tw.div`
-  flex flex-col
-  bg-white rounded-md
-  w-[170px] h-[165px]
-  cursor-pointer
-  text-gray-800
-  relative
-`;
-const Header = tw.div`
-  flex flex-row justify-between items-center
-  mb-2 mx-[14px] mt-[10px]
-  
-`;
-const Title = tw.div`
-  text-[18px] font-medium
-  
-`;
-
-const Main = tw.div`
-`;
-
-const Content = tw.div`
-  overflow-hidden
-  text-[15px]
-  mx-[14px]
-  
-`;
-
-const Date = tw.div`
-  text-[10px] font-light
-  flex justify-end
-  mb-0.5 mr-[2px]
-
-`;
-const Divider = styled.div`
-  height: 1px;
-
-  margin: 0 auto;
-  background-color: #898989;
-`;
-const Footer = tw.div`
-  w-full
-  flex flex-col
-  text-[15px]
-  px-[10px] mb-[10px] mt-0.5 
-  absolute bottom-0 left-0 right-0
-`;
-const FooterCol = tw.div`
-  flex flex-row justify-between
-  px-[2px]
-  
-`;
-const Category = tw.div`
-  font-medium
-`;
+import {
+  Card,
+  Header,
+  Title,
+  Main,
+  Content,
+  Date,
+  Divider,
+  Footer,
+  FooterCol,
+  Category,
+} from "./BulletinCard.styles";
 
 interface BulletinCardProps {
   id: string;
