+++ src/components/board/BulletinCardModal/BulletinCardModal.tsx
@@ -1,72 +1,23 @@
 import { useState, useEffect } from "react";
-import styled from "styled-components";
-import tw from "tailwind-styled-components";
+
 import axiosRequest from "@/api/index";
 import { ResData, Board } from "@/@types/index";
-import HeartBtn from "@/components/board/HeartBtn";
+import HeartBtn from "@/components/board/Button/HeartBtn/HeartBtn";
 
-const CardModalContainer = tw.div`
-  h-full
-  fixed z-10 inset-0
-  bg-black-semi-transparent
-  flex justify-center items-center
-  `;
-const CardModal = tw.div`
-  w-80 h-3/4 rounded-3xl
-  bg-white opacity-100
-  flex flex-col
-  relative
-  `;
-const Header = tw.div`
-  flex flex-row justify-center items-center
-  relative
-  mb-4
-  
-  `;
-const Category = tw.div`
-  text-2xl font-bold
-  my-3.5
-  `;
-const CloseBtn = tw.button`
-  absolute top-3.5 right-3.5
-  `;
-const Main = tw.div`
-  flex flex-col
-  px-6
-`;
-const Title = tw.div`
-  text-xl font-medium
-  mb-3.5
-`;
-const Content = tw.div`
-  text-base
-  overflow-auto
-  flex-grow
-
-`;
-const Divider = styled.div`
-  height: 1px;
-  width: 90%;
-  margin: 0 auto;
-  background-color: #e1e1e1;
-`;
-const FooterWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin: 8px 0px 16px;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-`;
-const Footer = tw.div`
-  flex flex-row justify-between items-center
-  w-full
-  px-6 pt-[10px]
-`;
-const Date = tw.div`
-  font-extralight
-`;
+import {
+  CardModalContainer,
+  CardModal,
+  Header,
+  Category,
+  CloseBtn,
+  Main,
+  Title,
+  Content,
+  Divider,
+  FooterWrap,
+  Footer,
+  Date,
+} from "./BulletinCardModal.styles";
 
 interface BulletinCardModalProps {
   closeModal: () => void;
