+++ src/components/board/Button/ChangeMbtiBtn/ChangeMbtiBtn.styles.ts
@@ -1,0 +1,5 @@
+import tw from "tailwind-styled-components";
+
+export const ChangeMbti = tw.div`
+  cursor-pointer
+`;
