+++ src/components/board/Button/HeartBtn/HeartBtn.styles.ts
@@ -1,0 +1,12 @@
+import tw from "tailwind-styled-components";
+
+export const Heart = tw.div`
+  flex flex-row items-center
+`;
+export const HeartIcon = tw.button`
+`;
+
+export const HeartCount = tw.div`
+  font-light
+  ml-[4px]
+`;
