+++ src/components/board/Button/PostBtn/PostBtn.styles.ts
@@ -1,0 +1,5 @@
+import tw from "tailwind-styled-components";
+
+export const Post = tw.div`
+  cursor-pointer
+`;
