+++ src/components/board/MbtiColorChip/MbtiColorChip.styles.ts
@@ -1,0 +1,5 @@
+import tw from "tailwind-styled-components";
+
+export const MbtiColor = tw.div`
+ flex flex-row gap-2
+`;
